extract out page logging function and extracting order details from pronto
